Split OpenVPN tunnel-state tests to separate file
