+++ desktop/packages/mullvad-vpn/test/e2e/installed/state-dependent/tunnel-state.spec.ts
@@ -95,61 +95,6 @@ test('App should connect with Shadowsocks', async () => {
   await expectConnected(page);
 });
 
-test('App should show correct tunnel protocol', async () => {
-  const tunnelProtocol = page.getByTestId('tunnel-protocol');
-  await expect(tunnelProtocol).toHaveText('WireGuard');
-
-  await exec('mullvad relay set tunnel-protocol openvpn');
-  await exec('mullvad relay set location se');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(tunnelProtocol).toHaveText('OpenVPN');
-});
-
-test('App should show correct OpenVPN transport protocol and port', async () => {
-  const inData = page.getByTestId('in-ip');
-
-  await expect(inData).toContainText(new RegExp(':[0-9]+'));
-  await expect(inData).toContainText(new RegExp('(TCP|UDP)$'));
-  await exec('mullvad relay set tunnel openvpn --transport-protocol udp --port 1195');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(inData).toContainText(new RegExp(':1195'));
-
-  await exec('mullvad relay set tunnel openvpn --transport-protocol udp --port 1300');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(inData).toContainText(new RegExp(':1300'));
-
-  await exec('mullvad relay set tunnel openvpn --transport-protocol tcp --port any');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(inData).toContainText(new RegExp(':[0-9]+'));
-  await expect(inData).toContainText(new RegExp('TCP$'));
-
-  await exec('mullvad relay set tunnel openvpn --transport-protocol tcp --port 80');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(inData).toContainText(new RegExp(':80'));
-
-  await exec('mullvad relay set tunnel openvpn --transport-protocol tcp --port 443');
-  await expectConnected(page);
-  await page.getByTestId('connection-panel-chevron').click();
-  await expect(inData).toContainText(new RegExp(':443'));
-
-  await exec('mullvad relay set tunnel openvpn --transport-protocol any');
-});
-
-test('App should show bridge mode', async () => {
-  await exec('mullvad bridge set state on');
-  await expectConnected(page);
-  const relay = page.getByTestId('hostname-line');
-  await expect(relay).toHaveText(new RegExp(' via ', 'i'));
-  await exec('mullvad bridge set state off');
-
-  await exec('mullvad relay set tunnel-protocol wireguard');
-});
-
 test('App should enter blocked state', async () => {
   await exec('mullvad debug block-connection');
   await expectError(page);
